SJTUsoso/requirements.txt
absl-py==0.8.0
asgiref==3.2.7
astor==0.8.0
certifi==2020.4.5.1
chardet==3.0.4
construct==2.5.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-formtools==2.2
django-haystack==2.8.1
django-import-export==2.1.0
django-js-asset==1.2.2
django-markdownx==3.0.1
django-mdeditor==0.1.18
django-php==0.1
django-reversion==3.0.7
django-xadmin==0.5.0
dom==0.9
et-xmlfile==1.0.1
future==0.18.2
gast==0.2.2
google-pasta==0.1.7
gprof2dot==2019.11.30
graphviz==0.13.2
grpcio==1.24.1
h5py==2.10.0
httplib2==0.18.0
idna==2.9
importlib-metadata==1.6.0
jdcal==1.4.1
jieba==0.42.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.2
MarkupPy==1.14
meteo-downloader==1.0
mysqlclient==1.4.6
numpy==1.17.2
odfpy==1.4.1
openpyxl==3.0.3
opt-einsum==3.1.0
pefile==2019.4.18
Pillow==7.1.1
protobuf==3.10.0
Pygments==2.6.1
PyPDF2==1.26.0
python-ptrace==0.9.5
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.42
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==0.16.0
Whoosh==2.7.4
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
